requirements.txt
# ---------------------------------------------------------------------------
# Farz — pinned dependencies
#
# Install: pip install -r requirements.txt
# Python 3.13 required
# ---------------------------------------------------------------------------

# API framework
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.8.0

# LLM
# Note: anthropic must be >=0.84.0 to be compatible with instructor 1.7.x
anthropic==0.84.0
instructor==1.7.2

# Database
supabase==2.15.1
psycopg2-binary==2.9.11

# Auth / JWT
python-jose[cryptography]==3.3.0

# Task queue
celery==5.4.0
redis==5.2.1

# Transcription
deepgram-sdk==3.9.0

# HTTP client
httpx==0.28.1

# Config / environment
python-dotenv==1.0.1

# ---------------------------------------------------------------------------
# Testing & QA
# ---------------------------------------------------------------------------
pytest==9.0.2
pytest-anyio==0.0.0

# ---------------------------------------------------------------------------
# Linting & type-checking
# ---------------------------------------------------------------------------
ruff==0.15.5
mypy==1.19.1
